+++ src/main/java/com/lssinni/AISM/controller/ReportController.java
@@ -1,7 +1,7 @@
 package com.lssinni.AISM.controller;
 
 import com.lssinni.AISM.domain.IamUser;
-import com.lssinni.AISM.service.IamService;
+import com.lssinni.AISM.service.ReportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,17 +9,17 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-public class IamController {
+public class ReportController {
 
-    private final IamService iamService;
+    private final ReportService iamService;
 
     @Autowired
-    public IamController(IamService iamService) {
+    public ReportController(ReportService iamService) {
         this.iamService = iamService;
     }
 
     // HTTP GET 요청으로 보고서를 생성하고 반환하는 엔드포인트
-    @GetMapping("/generate-credential-report")
+    @GetMapping("/generate")
     public List<IamUser> generateCredentialReport() {
         try {
             // IAMService를 통해 Credential Report 생성
